moved external dependencies into a shared subpackage
