+++ course_attendance_service/course_attendance_service/workflows/external/console.py
