+++ course_attendance_service/course_attendance_service/workflows/make_course_certificate/adapters/certificate_repo_filesystem.py
@@ -5,7 +5,7 @@ from pathlib import Path
 
 from ..core.certificate_repo import CertificateRepo
 from ..core.certificate_builder import WritableData
-from ..external.filesystem import Filesystem
+from ...external.filesystem import Filesystem
 
 
 @dataclass(frozen=True)
