+++ course_attendance_service/course_attendance_service/workflows/make_course_certificate/adapters/workshop_repo_zoom_test.py
@@ -5,7 +5,7 @@ from textwrap import dedent
 from unittest.mock import Mock
 
 from ..core.workflow import SessionRecord, WorkshopRecord
-from ..external.zoom_api import ZoomRestApi
+from ...external.zoom_api import ZoomRestApi
 from .workshop_repo_zoom import ZoomWorkshopRepo
