+++ course_attendance_service/course_attendance_service/workflows/make_course_certificate/workflows_test.py
@@ -6,7 +6,7 @@ from .core.workflow import PlannedWorkshopWorkflow
 from .core.workshop_repo import WorkshopRepo
 from .adapters.certificate_builder_console import ConsoleCertificateBuilder
 from .adapters.certificate_repo_filesystem import FilesystemCertificateRepo
-from .external.filesystem import Filesystem
+from ..external.filesystem import Filesystem
