+++ course_attendance_service/course_attendance_service/workflows/show_workshop_ids/adapters/console_printer.py
@@ -2,7 +2,7 @@ from dataclasses import dataclass
 from typing import List
 
 from ..core.workshop_list_presenter import WorkshopListPresenter, WorkshopData
-from ..external.console import Console
+from ...external.console import Console
 
 @dataclass(frozen=True)
 class ConsoleWorkshopListPresenter(WorkshopListPresenter):
