+++ course_attendance_service/course_attendance_service/workflows/show_workshop_ids/adapters/test_repo_zoom.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 from unittest.mock import Mock
 
 from .repo_zoom import ZoomWorkshopRepo
-from ..external.zoom_api import ZoomRestApi
+from ...external.zoom_api import ZoomRestApi
 
 
 def test_repo_can_get_list_of_workshops_from_zoom_api():
